rechecked all core tests, activated missed ones, modified partly to work, the env test checks now all 3 data paths

+++ core/patterndetection/circlePatternGenerator.cpp
@@ -128,9 +128,9 @@ int CirclePatternGenerator::getBestToken(DpImage &query, double &score, corecvs:
                 std::cout << corscore << std::endl;
                 if (corscore < maxC && -corscore > corrThreshold)
                 {
-                    maxC = corscore;
-                    maxIdx = id;
-                    maxRI = idp;
+                    maxC   = corscore;
+                    maxIdx = (int)id;
+                    maxRI  = idp;
                 }
             }
         }
